Introduce Amazon Block Storage  Manager (EBS)
This commit sets up the core structure of the new AWS EC2 block storage
manager. It provides just the basic stubs and tests that are to be
extended to allow for complete management of Amazon block storage

The refresher collects data about existing volumes and private
snapshots. Volumes are not yet linked to disks of virtual machines
collected by the cloud manager.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>