Rename Amazon BlockStorageManager to StorageManager::Ebs
Initial implementation of the block storage manager presumed that Amazon
provider would only have a single block storage manager and consequently
used the generic class name `Amazon::BlockStorageManager`. @rolivery
suggested to change the name to indicate the actual type of the manager,
which in this case is EBS.

This patch thus replaces all occurrences where `BlockStorageManager` was
used before with `StorageManager::Ebs`.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>