Add links between cloud volumes and base snapshots
This patch adds the ability to identify base snapshot, i.e. the snapshot
that was used to create a specific volume. Both legacy parser and
inventory object parser are updated, along with the spec. The latter
adds tests for specific volume and snapshot to make sure proper values
are identified.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>