Link cloud volumes to corresponding VM disks
This patch updates the EBS storage manager with the ability to
cross-link attached cloud volumes with corresponding VM disks. Since the
cloud manager already creates disk entries based on the device block
mappings that are available to the insntace, EBS storage manager just
updates the disk with it's size and creates a backing link. Legacy
refresh parser as well as the new inventory object are extended to
support the cross-linking.

This patch also adds stubbed data mocking an attachment for a specific
cloud volume and verifies that the EBS storage manager properly updates
the existing disk.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>