Collect block device mappings in AWS cloud manager
Besides collecting ephemeral storage in the AWS cloud manager refresh
parser, we are adding inventory of attached block devices. This is
necessary because with the introduction of the EBS manager we are adding
disks there as well. If the cloud manager does not collect attached
block devices, those disks collected by the EBS manager will be removed
from the inventory.

The suggested change does not require any change in the core repository.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>