Remove `require_nested :Runner` from two invalid models
Both the Amazon::NetworkManager::EventCatcher and
Amazon::NetworkManager::MetricsCollectorWorker don't have corresponding
nested event_catcher/runner.rb and metrics_collector_worker/runner.rb
directories respectively, thus `require_nested` fails.

The reason this doesn't seem to be a problem is we don't eager load in
manageiq, and it is possible that these two classes are currently not in
use.