If VM was scheduled for refresh, always refresh the references
If VM was scheduled for refresh, always refresh the references that
we can obtain by scanning the Vm in API and our DB. This way, we
should always get a Vm that has all references filled.

The alternative for this would be a multiphase refresh, where each
refresh can be planning another refresh for missing references.

In this case, we output a fully refreshed Vms in one step, which
will be later configurable, as this is handy for integration
with external tools. E.g. when we need to send refreshed data
to ansible, we need Vm completelly refreshed after 1 refresh, not
possible many incremental refreshes.