Connect to correct region with S3
Current code always connects to S3's region that of cloud manager.
This works fine when buckets that we're accessing are within
that region, but results in error when accessing bucket from
any other region.

With this commit we update provider_object for CloudObjectStoreContainer
and CloudObjectStoreObject so that it now connects to appropriate region.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>