Remap :name parameter into :bucket parameter
Current implementation assumes that frontend composes
all the options that backend then just sends to AWS. But
there is an exception with :name option. While it is
needed on frontend (to display flash), it is rejected by AWS.

Initial idea was to just remove the :name option here in backend,
but turns out that renaming it to :bucket works better. Hence this
commit.

Signed-off-by: Miha Pleško <miha.plesko@xlab.si>