Handle SSA Response from AWS Scanning Instance
Previously the request to perform SSA on AWS by the Scanning Instance was implemented.
We are now implementing the code to process the response from the the SSA scan upon
completion.

Note that much of the code to do this, in perform_metadata_sync,
was taken from the ScanningMixin method scan_via_miq_vm.  The intent will be to
mitigate duplication when possible moving forward but for now we wanted to get
this running at the very least.

Also note that for now this code will be reading the first available response from
SQS.  If more than one SSA request is issued simultaneously there is a good chance
that the responses will be handled out of order.  We intend to implement a worker
to perform this task in the near future to handle this issue.