Adjust test result to match new data_type
The security_groups key now defaults to a data_type of :array_integer.  The resulting value is an Array of Integers
as opposed to an array with the first value being an integer and the second being a human readable value.  This fix modifies the resulting
hash to expect an array of Integers.