Amazon SSA Response Handling Thread
Add the Amazon SSA response thread which will be integrated with the
AgentCoordinatorWorker.  This thread checks the SQS queues set up
by the Amazon SSA Agents to see if any SSA responses have been added.
If so the responses are pulled out of the appropriate S3 buckets and
processed.  The appropriate Job is then updated with the status.