Raise Exception if AMI not visible
In the case that the AMI in AWS that is to be used for Smartstate is not visible
to the user, raise a meaningful error which will be visible in the log.

In a separate PR we will look at marking any jobs waiting for the agent to be deployed
as failed when this happens.