Add patch to aws-sdk-core to fix auth bug
There is a bug with the aws-sdk that prevents the API from working
properly when using a proxy, and the auth for the proxy includes a
special character (like a question mark: '?').  This is described here:

https://github.com/aws/aws-sdk-ruby/pull/1760

Since the release of the fix is unknown, this currently monkey patches
the fix to MIQ to allow it to work without it.