Add patch to original code comparison specs
This spec is a bit complicated, but effectively it is doing a diff of
the un-changed parts of the monkey patched code from the `aws-sdk-core`
repo, and making sure they are still the same.

This required a minor update to the patch to basically extract the
current source location prior to applying the patch, but other than
that, this should function as expected.  The approach taken avoided any
RubyLexing, and opted for a simpler file reading approach since there
was known quantities for the changed code v.s. the original.