Remove allow_delete logic
This PR is reverting https://github.com/ManageIQ/manageiq-providers-amazon/pull/457 as that is causing an issue in deleting auth keys created, without a public key.

Also pr#457 doesn't even seem to solve the issue of the bug it was suposed to be fixing.

This PR fixes BZ: https://bugzilla.redhat.com/show_bug.cgi?id=1718833