Fix EKS scheduled auth check returning missing creds
While we are using a bearer token to authenticate with k8s API, we
aren't storing this token in the database.  We are only storing userid
and password in the authentication record, with an authtype of "bearer"
this was causing the scheduled authentication_check to fail with
"missing credentials" because it didn't see anything in the `:bearer`
column.