Fix AgentCoordinatorWorker sync_workers
The AgentCoordinatorWorker runs 1 copy if there are any valid amazon
providers in the zone and the smartproxy role is enabled.

It was pretending to be a PerEmsWorker and overriding the
desired_queue_names method but didn't override the instance queue_name
so sync_workers was constantly getting:
```
ManageIQ::Providers::Amazon::AgentCoordinatorWorker.sync_workers
Workers are being synchronized: Current: [\"ems_0\"], Desired:
[\"ems_agent_coordinator\"]"
```

Instead of trying to be a PerEmsWorker simply make it a standard worker
that checks for any valid EMSs before saying it has the required_role.