Use Vmdb::Utils.http_proxy_uri when adding AWS provider
The typical method for setting a proxy is via the
`Settings.http_proxy.ec2` setting which is accessed via
the `Vmdb::Utils.http_proxy_uri` method.

This is used during the instance connect/verify methods but not during
the class connect/verify methods.  This leads to the inability to add an
AWS provider if a proxy is required to access it.