Refactor regions spec to not hard-code atypical regions
The regions spec has a hard-coded list of "atypical regions" which
really just means regions which exist but are not available to us via
the aws-sdk `describe_regions` API call.

It would be simpler to just check that all of the regions that we do
have access to are in the regions list.  This would allow us to update
the `config/regions.yml` list in an automated way.